requirements.txt
python-telegram-bot>=20.0
ccxt>=4.0.0
python-dotenv>=1.0.0
flask